Revert moving packages
